Fix int32 overflow in integer fast paths at the int.MinValue/MaxValue boundary
Integer-tagged arithmetic fast paths performed 32-bit operations that wrap
or raise hardware overflow exceptions at the int32 boundary:

- -((1<<31)|0) wrapped to -2147483648 instead of 2147483648
- ((1<<31)|0) / -1 and % -1 threw "Arithmetic operation resulted in an
  overflow" (CLR ArithmeticException escaping to the caller)
- 2147483647++ wrapped to -2147483648; ((1<<31)|0)-- and -= 1 wrapped
  to 2147483647

Widen to 64-bit math before the operation, matching the existing pattern
already used by the +, - and * binary operator fast paths.

+++ Jint/Runtime/Interpreter/Expressions/JintAssignmentExpression.cs
@@ -154,7 +154,7 @@ internal sealed class JintAssignmentExpression : JintExpression
 
                         if (AreIntegerOperands(originalLeftValue, rval))
                         {
-                            newLeftValue = JsNumber.Create(originalLeftValue.AsInteger() - rval.AsInteger());
+                            newLeftValue = JsNumber.Create((long) originalLeftValue.AsInteger() - rval.AsInteger());
                         }
                         else if (JintBinaryExpression.AreNonBigIntOperands(originalLeftValue, rval))
                         {
